+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Release NEXT
 ------------
 - Allow user to define his token lifetime.
 - On update, instance is not returned in the old state anymore.
+- Move throttling tasks from openstack plugin into structure.
 
 Release 0.122.0
 ---------------